CHROMIUM: sched: tune: allow nested schedtune groups
Let's remove restriction that schedtune groups can not be nested. Note
that they are still completely independent even if they are nested, but
at least we can build proper hierarchies now.

We also extend maximum number of schedtune groups to 16 for now.

CHROMIUM Note: Apparently the plan is to drop schedtune in favor of
utilization clamping. Future kernel rebases should hopefully be able
to drop this patch once utilization clamping is accepted upstream and
we've moved to it. Check out the second bug in the BUG= field.

BUG=b:74017669 b:116869240
TEST=Build and boot Cheza with EAS enabled

Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/425443
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Change-Id: Icdf8db99dc2d252b782b228caedefa6abf41a9a6
Reviewed-on: https://chromium-review.googlesource.com/1235093
Commit-Ready: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>